Refactor search client functions
