+++ tests/api/data-validation.test.js
@@ -56,9 +56,6 @@ describe('[P2][Sev2] Search API: Validate data in index', () => {
       
           // Create a route to access the Search API.
           searchApiRoute = await getSearchApiRoute()
-
-          // Disable SSL validation so we can connect to the search-api route.
-          process.env.NODE_TLS_REJECT_UNAUTHORIZED = 0
         })
         resourceList.forEach((resource) => {
           // There can be multiple occurrences of the same resource kind with different API groups; therefore
