+++ tests/common-lib/searchClient.js
@@ -5,10 +5,8 @@
  */
 const { fail } = require('assert')
 const { performance } = require('perf_hooks')    
-const {
-    formatResourcesFromSearch,
-    formatFilters } = require('./index')
 const request = require('supertest')
+const lodash = require('lodash')
 
 /**
  * Query the Search API using the given filters.
@@ -22,28 +20,11 @@ async function getResourcesFromSearch(kind,
     apigroup,
     namespace = '--all-namespaces',
     cluster = { type: 'hub', name: 'local-cluster' }){
+    // Build the search api query.
     const filters = formatFilters(kind, apigroup, namespace, cluster)
-  
+    const query = searchQueryBuilder({ filters })
     // Fetch data from the search api.
-    var query = searchQueryBuilder({ filters })
-  
-    // Monitor how long search took to return results.
-    var startTime = performance.now()
-    var resp = await sendRequest(query, token)
-    var endTime = performance.now()
-    var totalElapsedTime = endTime - startTime
-  
-    if (totalElapsedTime > 30000) {
-      fail(
-        `Search required more than 30 seconds to return resources for ${kind}. (TotalElapsedTime: ${totalElapsedTime})`
-      )
-    } else if (totalElapsedTime > 1000) {
-      console.warn(
-        `Search required more than 1 second to return resources for ${kind}. (TotalElapsedTime: ${totalElapsedTime.toFixed(
-          2
-        )})`
-      )
-    }
+    const resp = await sendRequest(query, token)
   
     return formatResourcesFromSearch(resp)
   }
@@ -83,13 +64,83 @@ async function getResourcesFromSearch(kind,
  * @param {object} options Additional options for sending the request.
  * @returns
  */
- function sendRequest(query, token, options = {}) {
+function sendRequest(query, token, options = {}) {
+    process.env.NODE_TLS_REJECT_UNAUTHORIZED = 0 // Disable SSL validation so we can connect to the search-api route.
+    
+    // Monitor how long search took to return results.
+    const startTime = performance.now()
+     
     return request(searchApiRoute)
       .post('/searchapi/graphql')
       .send(query)
       .set({ Authorization: `Bearer ${token}` })
       .expect(200)
-  }
+      .then((r) => {
+        const endTime = performance.now()
+        const totalElapsedTime = endTime - startTime
+        
+        if (totalElapsedTime > 10000) {
+            fail(
+                `Search required more than 10 seconds to return resources for query [${query}]. (TotalElapsedTime: ${totalElapsedTime})`
+            )
+        } else if (totalElapsedTime > 1000) {
+            console.warn(
+                `Search required more than 1 second to return resources for query [${query}]. (TotalElapsedTime: ${totalElapsedTime.toFixed(2)})`
+            )
+        } 
+        
+        return r   
+      })
+}
+
+/**
+ * Format resources for search queries.
+ * @param {*} resources A list of resources that will be formated as an object containing name and namespace.
+ * @returns `formatedResources` Formatted array of resource object.
+ */
+function formatResourcesFromSearch(resources) {
+    return lodash.get(resources, 'body.data.searchResult[0].items')
+        .filter((items) => items.namespace) // We're only interested in resources that have a namespace.
+        .map((item) => ({
+            cluster: item.cluster,
+            kind: item.kind,
+            name: item.name,
+            namespace: item.namespace,
+        }))
+}
+
+/**
+ * Format filters for search queries.
+ * @param {string} kind The kind of resource to filter.
+ * @param {Object} group The API group to filter the resources against.
+ * @param {string} namespace The namespace to filter the resources against.
+ * @param {Object} cluster The cluster to filter the resources against.
+ * @returns `filter` Formatted array of object filters.
+ */
+ function formatFilters(
+    kind,
+    group,
+    namespace = '--all-namespaces',
+    cluster = { type: 'hub', name: 'local-cluster' }
+  ) {
+    const filter = []
+  
+    // Add namespace filter
+    if (namespace !== '--all-namespaces')
+      filter.push({ property: 'namespace', values: [namespace] })
+  
+    // Add kind filter
+    filter.push({ property: 'kind', values: [kind] })
+  
+    // Add group filter
+    if (group.useAPIGroup && group.name != 'v1')
+      filter.push({ property: 'apigroup', values: [group.name] })
+  
+    // Add cluster filter
+    filter.push({ property: 'cluster', values: [cluster.name] })
+  
+    return filter
+}
   
 exports.getResourcesFromSearch = getResourcesFromSearch
 exports.searchQueryBuilder = searchQueryBuilder
